P4-3091 moving controller contraints into the controller level package.
