+++ src/main/kotlin/uk/gov/justice/digital/hmpps/pecs/jpc/controller/constraints/JourneySearchConstraintValidator.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.pecs.jpc.constraint
+package uk.gov.justice.digital.hmpps.pecs.jpc.controller.constraints
 
 import uk.gov.justice.digital.hmpps.pecs.jpc.controller.HtmlController
 import javax.validation.ConstraintValidator
